move Input and Span to span.py
